Re-organize components
