+++ apps/client/package.json
@@ -30,6 +30,7 @@
     "@radix-ui/react-tooltip": "^1.2.7",
     "@tailwindcss/typography": "^0.5.16",
     "@tailwindcss/vite": "^4.0.6",
+    "@tanstack/react-router": "^1.131.27",
     "@tauri-apps/api": "^2",
     "@tauri-apps/plugin-clipboard-manager": "~2.3.0",
     "@tauri-apps/plugin-deep-link": "~2.4.0",
