+++ apps/client/src/components/core/PublicKeyAvatar.tsx
