+++ apps/client/src/components/core/ValueTable.tsx
