+++ apps/client/src/components/frogcrypto/FrogViewer.tsx
@@ -1,23 +1,23 @@
-import { useEffect, useState } from "react";
-import { useFrogCrypto } from "../lib/store";
-import { Card, CardContent } from "./ui/card";
-import { ScrollArea } from "./ui/scroll-area";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent } from "@/components/ui/card";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import { Switch } from "@/components/ui/switch";
 import {
-  requestFrog,
-  requestScore,
-  listFrogs,
+  FrogData,
   FrogPod,
-  FrogData
+  listFrogs,
+  requestFrog,
+  requestScore
 } from "@/lib/rpc";
-import { toast } from "sonner";
+import { useFrogCrypto } from "@/lib/store";
 import {
   Collapsible,
   CollapsibleContent,
   CollapsibleTrigger
 } from "@radix-ui/react-collapsible";
-import { Switch } from "./ui/switch";
-import { listen, emit } from "@tauri-apps/api/event";
+import { emit, listen } from "@tauri-apps/api/event";
+import { useEffect, useState } from "react";
+import { toast } from "sonner";
 import { RARITY_SHADOW_COLORS } from "./FrogCrypto";
 
 interface FrogViewerProps {
