+++ apps/client/src/components/settings/DebugView.tsx
@@ -12,10 +12,10 @@ import {
 } from "lucide-react";
 import { useEffect, useState } from "react";
 import { toast } from "sonner";
-import { testDeepLink } from "../lib/deeplink";
-import { Button } from "./ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
-import { Input } from "./ui/input";
+import { testDeepLink } from "../../lib/deeplink";
+import { Button } from "../ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
+import { Input } from "../ui/input";
 
 interface AppConfig {
   features: {
